+++ django/applications/catmaid/static/js/widgets/neuron_navigator.js
@@ -4,7 +4,6 @@
   CATMAID
   AnalyzeArbor,
   Arbor,
-  ConnectorTable,
   InstanceRegistry,
   NeuronDendrogram,
   project,
@@ -2085,8 +2084,8 @@
 
     connectortable_button.onclick = (function() {
       if (this.skeleton_ids.length > 0) {
-        var TNT = new CATMAID.ConnectorTable(this.skeleton_ids[0]);
-        WindowMaker.create('connector-table', TNT);
+        WindowMaker.create('connector-table',
+            this.getSelectedSkeletonModels());
       }
     }).bind(this);