+++ django/applications/catmaid/static/js/widgets/table-connector.js
@@ -5,247 +5,33 @@
 
   "use strict";
 
-  var ConnectorTable = function(optionalSkid)
+  /**
+   * Create a new connector table, optional with a set of initial skeleton
+   * models.
+   */
+  var ConnectorTable = function(skeletonModels)
   {
     this.widgetID = this.registerInstance();
 
-    /** Pointer to the existing instance of table. */
+    // This skeleton source takes care of internal skeleton management. It is
+    // not registered. It is the input skeleton sink, but the output is handled
+    // with a second soucre
+    var update = this.update.bind(this);
+    this.skeletonSource = new CATMAID.BasicSkeletonSource(this.getName() + " (input)", {
+      register: false,
+      handleAddedModels: update,
+      handleChangedModels: update,
+      handleRemovedModels: update
+    });
+    // A skeleton source to collect results in
+    this.resultSkeletonSource = new CATMAID.BasicSkeletonSource(this.getName());
+
+    // The displayed data table
     this.connectorTable = null;
 
-    var self = this;
-    var asInitValsSyn = [];
-    var skeletonID = optionalSkid ? optionalSkid : -1;
-    var possibleLengths = CATMAID.pageLengthOptions;
-    var possibleLengthsLabels = CATMAID.pageLengthLabels;
-
-    this.updateConnectorTable = function() {
-      self.setSkeleton( -1 );
-      self.connectorTable.fnClearTable( 0 );
-      self.connectorTable.fnDraw();
-    };
-
-    this.refreshConnectorTable = function() {
-      self.connectorTable.fnClearTable( 0 );
-      self.connectorTable.fnDraw();
-    };
-
-    this.setSkeleton = function( skeleton_id ) {
-      skeletonID = skeleton_id;
-    };
-
-    this.getSkeletonId = function() {
-      return skeletonID;
-    };
-
-    this.init = function (pid) {
-      var widgetID = this.widgetID;
-      var tableid = '#connectortable' + widgetID;
-
-      self.connectorTable = $(tableid).dataTable(
-        {
-          "bDestroy": true,
-          "sDom": '<"H"lr>t<"F"ip>',
-          "bProcessing": true,
-          "bServerSide": true,
-          "bAutoWidth": false,
-          "iDisplayLength": possibleLengths[0],
-          "sAjaxSource": django_url + project.id + '/connector/table/list',
-          "fnServerData": function (sSource, aoData, fnCallback) {
-
-            if( skeletonID === -1 ) {
-              skeletonID = SkeletonAnnotations.getActiveSkeletonId();
-            }
-
-            if (!skeletonID) {
-              CATMAID.msg('BEWARE', 'You need to activate a treenode to display ' +
-                  'the connector table of its skeleton.');
-              skeletonID = 0;
-            }
-
-            aoData.push({
-              "name": "relation_type",
-              "value" : $('#connector_relation_type' + widgetID + ' :selected').attr("value")
-            });
-            aoData.push({
-              "name" : "pid",
-              "value" : pid
-            });
-            aoData.push({
-              "name" : "skeleton_id",
-              "value" : skeletonID
-            });
-
-            $.ajax({
-              "dataType": 'json',
-              "cache": false,
-              "type": "POST",
-              "url": sSource,
-              "data": aoData,
-              "success": fnCallback
-            });
-
-          },
-          "aLengthMenu": [
-            possibleLengths,
-            possibleLengthsLabels
-          ],
-          "bJQueryUI": true,
-          "aoColumns": [
-            {
-              "bSearchable": false,
-              "bSortable": true
-            }, // connector id
-            {
-              "sClass": "center",
-              "bSearchable": false
-            }, // other skeleton id
-            {
-              "sClass": "center",
-              "bSearchable": false
-            }, // x
-            {
-              "sClass": "center",
-              "bSearchable": false
-            }, // y
-            {
-              "sClass": "center",
-              "bSearchable": false
-            }, // z
-            {
-              "sClass": "center",
-              "bSearchable": false,
-              "bSortable": true,
-              data: 4,
-              render: function(data, type, row, meta) {
-                return project.focusedStackViewer.primaryStack.projectToStackZ(row[4], row[3], row[2]);
-              },
-            }, // section index
-            {
-              "bSearchable": false,
-              "bSortable": true,
-              data: 5,
-            }, // connectortags
-            {
-              "bSearchable": false,
-              "bSortable": true,
-              data: 6,
-            }, // confidence
-            {
-              "bSearchable": false,
-              "bSortable": true,
-              data: 7,
-            }, // target confidence
-            {
-              "bSearchable": false,
-              "bSortable": true,
-              data: 8,
-            }, // number of nodes
-            {
-              "bVisible": true,
-              "bSortable": true,
-              data: 9,
-            }, // username
-            {
-              "bSearchable": false,
-              "bSortable": true,
-              "bVisible": true,
-              data: 10,
-            }, // treenodes
-            {
-              "bSearchable": false,
-              "bSortable": true,
-              "bVisible": true,
-              data: 11,
-              render: function(data, type, row, meta) {
-                var d = new Date(data);
-                return d.getFullYear() + '-' + (d.getMonth() + 1) + '-' + d.getDate()
-                    + ' ' + d.getHours() + ':' + d.getMinutes();
-              }
-            } // last modified
-          ]
-        });
-
-      $(tableid + " tfoot input").keyup(function () { /* Filter on the column (the index) of this element */
-        self.connectorTable.fnFilter(this.value, $("tfoot input").index(this));
-      });
-
-      /*
-       * Support functions to provide a little bit of 'user friendlyness' to the textboxes in
-       * the footer
-       */
-      $(tableid + " tfoot input").each(function (i) {
-        asInitValsSyn[i] = this.value;
-      });
-
-      $(tableid + " tfoot input").focus(function () {
-        if (this.className == "search_init") {
-          this.className = "";
-          this.value = "";
-        }
-      });
-
-      $(tableid + " tfoot input").blur(function (i) {
-        if (this.value == "") { // jshint ignore:line
-          this.className = "search_init";
-          this.value = asInitValsSyn[$("tfoot input").index(this)];
-        }
-      });
-
-      $(tableid + " tbody").on('dblclick', 'td', function () {
-        // Allow clicking on the connector ID (column 0) to select it rather
-        // than the target treenode.
-        if ($(this).index() !== 0) return;
-        var idToActivate = self.connectorTable.fnGetData(this);
-
-        SkeletonAnnotations.staticMoveToAndSelectNode(idToActivate);
-
-        return false;
-      });
-
-      $(tableid + " tbody").on('dblclick', 'tr', function () {
-        var idToActivate, skeletonID;
-        var aData = self.connectorTable.fnGetData(this);
-        // retrieve coordinates and moveTo
-        var x = parseFloat(aData[2]);
-        var y = parseFloat(aData[3]);
-        var z = parseFloat(aData[4]);
-
-        // If there is a partner treenode, activate that - otherwise
-        // activate the connector itself:
-        if (aData[10]) {
-          idToActivate = parseInt(aData[10], 10);
-          skeletonID = parseInt(aData[1], 10);
-        } else {
-          idToActivate = parseInt(aData[0], 10);
-          skeletonID = null;
-        }
-
-        SkeletonAnnotations.staticMoveTo(z, y, x,
-          function () {
-            SkeletonAnnotations.staticSelectNode(idToActivate, skeletonID);
-          });
-      });
-
-      $('#connector_relation_type' + widgetID).change(function() {
-        var numberOfNodesText, otherSkeletonText, otherTreenodeText, adjective;
-        self.connectorTable.fnDraw();
-        if ($('#connector_relation_type' + widgetID + ' :selected').attr("value") === "0") {
-          adjective = "source";
-        } else {
-          adjective = "target";
-        }
-        numberOfNodesText = "# nodes in " + adjective + " skeleton";
-        otherSkeletonText = adjective + " skeleton ID";
-        otherTreenodeText = adjective + " treenode ID";
-        $("#connector_nr_nodes_top" + widgetID).text(numberOfNodesText);
-        $("#connector_nr_nodes_bottom" + widgetID).text(numberOfNodesText);
-        $("#other_skeleton_top" + widgetID).text(otherSkeletonText);
-        $("#other_skeleton_bottom" + widgetID).text(otherSkeletonText);
-        $("#other_treenode_top" + widgetID).text(otherTreenodeText);
-        $("#other_treenode_bottom" + widgetID).text(otherTreenodeText);
-      });
-
-    };
+    if (skeletonModels) {
+      this.skeletonSource.append(skeletonModels);
+    }
   };
 
   ConnectorTable.prototype = {};
@@ -259,26 +45,234 @@
     this.unregisterInstance();
   };
 
+  ConnectorTable.prototype.getWidgetConfiguration = function() {
+    return {
+      controlsID: 'connector-table-controls',
+      createControls: function(controls) {
+        var self = this;
+        var sourceSelect = CATMAID.skeletonListSources.createSelect(this.skeletonSource);
+        controls.appendChild(sourceSelect);
+
+        var add = document.createElement('input');
+        add.setAttribute("type", "button");
+        add.setAttribute("value", "Add");
+        add.onclick = this.skeletonSource.loadSource.bind(this.skeletonSource);
+        controls.appendChild(add);
+
+        var clear = document.createElement('input');
+        clear.setAttribute("type", "button");
+        clear.setAttribute("value", "Clear");
+        clear.onclick = function() {
+          self.skeletonSource.clear();
+          self.update();
+        };
+        controls.appendChild(clear);
+
+        var add = document.createElement('input');
+        add.setAttribute("type", "button");
+        add.setAttribute("id", "refresh_connectortable_current_skeleton" + this.widgetID);
+        add.setAttribute("value", "Refresh");
+        add.onclick = this.update.bind(this);
+        controls.appendChild(add);
+
+        var relation = CATMAID.DOM.createSelect(
+          "connector_relation_type" + this.widgetID, [
+          {title: 'Incoming connectors', value: "postsynaptic_to"},
+          {title: 'Outgoing connectors', value: "presynaptic_to"},
+          {title: 'Gap junction connectors', value: "gapjunction_with"},
+          {title: 'Abutting connectors', value: "abutting"}],
+          "presynaptic_to");
+        relation.onchange = this.update.bind(this);
+        var relationLabel = document.createElement('label');
+        relationLabel.appendChild(document.createTextNode('Type'));
+        relationLabel.appendChild(relation);
+        controls.appendChild(relationLabel);
+
+        var exportCSV = document.createElement('input');
+        exportCSV.setAttribute("type", "button");
+        exportCSV.setAttribute("value", "Export CSV");
+        exportCSV.setAttribute("title", "Export a CSV file for the currently displayed table");
+        exportCSV.onclick = this.exportCSV.bind(this);
+        controls.appendChild(exportCSV);
+      },
+      contentID: ' connector-table-content',
+      createContent: function(content) {
+        var self = this;
+        var possibleLengths = CATMAID.pageLengthOptions;
+        var possibleLengthsLabels = CATMAID.pageLengthLabels;
+        var widgetID = this.widgetID;
+        var tableid = '#connectortable' + widgetID;
+
+        var container = document.createElement('div');
+        content.appendChild(container);
+
+        var table = document.createElement('table');
+        table.style.width = "100%";
+        content.appendChild(table);
+
+        this.connectorTable = $(table).DataTable({
+          dom: "lrphtip",
+          paging: true,
+          order: [],
+          serverSide: true,
+          lengthMenu: [CATMAID.pageLengthOptions, CATMAID.pageLengthLabels],
+          ajax: function(data, callback, settings) {
+            var relationType = $('#connector_relation_type' + widgetID + ' :selected').val();
+            var params = {
+              skeleton_ids: self.skeletonSource.getSelectedSkeletons(),
+              relation_type: relationType
+            };
+
+            // Locally resolve request if no skeletons are provided
+            if (0 === params.skeleton_ids.length) {
+              callback({
+                draw: data.draw,
+                recordsTotal: 0,
+                recordsFiltered: 0,
+                data: []
+              });
+              return;
+            }
+              
+            if (undefined !== data.length) {
+              params.range_start = data.start;
+              params.range_length = data.length;
+            }
+
+            if (data.order && data.order.length > 0) {
+              params.sort_column = data.order[0].column;
+              params.sort_dir = data.order[0].dir;
+
+              // Correct for artificial stack section column
+              if (params.sort_column > 4) {
+                params.sort_column = params.sort_column - 1;
+              }
+            }
+
+            CATMAID.fetch(project.id +  "/connectors/", "GET", params)
+              .then(function(result) {
+                // Populate table
+                callback({
+                  draw: data.draw,
+                  recordsTotal: result.total_count,
+                  recordsFiltered: result.total_count,
+                  data: result.links
+                });
+                // Populate result skeleton source
+                var models = result.links.reduce(function(o, link) {
+                  var skid = link[1];
+                  // TODO Color according to relation type for this link
+                  o[skid]  = new CATMAID.SkeletonModel(skid);
+                  return o;
+                }, {});
+                self.resultSkeletonSource.clear();
+                self.resultSkeletonSource.append(models);
+              })
+              .catch(CATMAID.handleError);
+          },
+          columns: [
+            {data: 0, className: "cm-center", title: "Connector"},
+            {data: 1, className: "cm-center", title: "Skeleton ID"},
+            {data: 2, className: "cm-center", title: "X"},
+            {data: 3, className: "cm-center", title: "Y"},
+            {data: 4, className: "cm-center", title: "Z"},
+            {
+              data: 4,
+              title: "S",
+              className: "cm-center",
+              render: function(data, type, row, meta) {
+                return project.focusedStackViewer.primaryStack.projectToStackZ(row[4], row[3], row[2]);
+              },
+            }, // section index
+            {data: 5, className: "cm-center", title: "Source confidence"},
+            {data: 6, className: "cm-center", title: "Partner confidence"},
+            {data: 7, className: "cm-center", title: "Tags"},
+            {data: 8, className: "cm-center", title: "# Skeleton nodes"},
+            {data: 9, className: "cm-center", title: "User"},
+            {data: 10, className: "cm-center", title: "Treenode ID"},
+            {
+              data: 11,
+              title: "Last modified",
+              className: "catmaid.center",
+              render: function(data, type, row, meta) {
+                var d = new Date(data);
+                return d.getFullYear() + '-' + (d.getMonth() + 1) + '-' + d.getDate()
+                    + ' ' + d.getHours() + ':' + d.getMinutes();
+              }
+            },
+          ]
+        }).on('dblclick', 'tbody td', function() {
+          var cell = self.connectorTable.cell( this );
+          if (!cell || cell[0][0].column !== 0) {
+            return;
+          }
+
+          var clickedId = cell.data();
+          SkeletonAnnotations.staticMoveToAndSelectNode(clickedId);
+
+          return false;
+        }).on('dblclick', 'tbody tr', function () {
+          var row = self.connectorTable.row(this);
+          var data = row.data();
+          // retrieve coordinates and moveTo
+          var x = parseFloat(data[2]);
+          var y = parseFloat(data[3]);
+          var z = parseFloat(data[4]);
+
+          // If there is a partner treenode, activate that - otherwise
+          // activate the connector itself:
+          var idToActivate, skeletonID;
+          if (data[10]) {
+            idToActivate = parseInt(data[10], 10);
+            skeletonID = parseInt(data[1], 10);
+          } else {
+            idToActivate = parseInt(data[0], 10);
+            skeletonID = null;
+          }
+
+          SkeletonAnnotations.staticMoveTo(z, y, x, function() {
+            SkeletonAnnotations.staticSelectNode(idToActivate, skeletonID);
+          });
+        });
+      }
+    };
+  };
+
   /**
    * Export the currently displayed table as CSV.
    */
   ConnectorTable.prototype.exportCSV = function() {
     if (!this.connectorTable) return;
     var relation = $('#connector_relation_type' + this.widgetID + ' :selected').val();
-    var table = this.connectorTable.DataTable();
-    var header = table.columns().header().map(function(h) {
+    var header = this.connectorTable.columns().header().map(function(h) {
       return $(h).text();
     });
-    var connectorRows = table.rows({"order": "current"}).data();
+    var connectorRows = this.connectorTable.rows({"order": "current"}).data();
     var csv = header.join(',') + '\n' + connectorRows.map(function(row) {
       return row.join(',');
     }).join('\n');
     var blob = new Blob([csv], {type: 'text/plain'});
+    var skeletonIds = this.skeletonSource.getSelectedSkeletons().join('-');
     saveAs(blob, "catmaid-connectors-" + relation + "-skeleton-" +
-        this.getSkeletonId() + ".csv");
+        skeletonIds + ".csv");
+  };
+
+  /**
+   * Update display based on current skeleton source content.
+   */
+  ConnectorTable.prototype.update = function() {
+    if (this.connectorTable) {
+      this.connectorTable.ajax.reload();
+    }
   };
 
   // Export widget
   CATMAID.ConnectorTable = ConnectorTable;
+
+  // Register widget with CATMAID
+  CATMAID.registerWidget({
+    key: 'connector-table',
+    creator: ConnectorTable
+  });
 
 })(CATMAID);